// frontend/src/types.interfaces.ts
export interface AggResults {
    SEX: string | null;
    AGE: number | null;
    CLASIFFICATION_FINAL: number;
    PNEUMONIA: string | null;
    PREGNANT: string | null;
    DIABETES: string | null;
    COPD: string | null;
    ASTHMA: string | null;
    INMSUPR: string | null;
    HIPERTENSION: string | null;
    CARDIOVASCULAR: string | null;
    OTHER_DISEASE: string | null;
    RENAL_CHRONIC: string | null;
    TOBACCO: string | null;
    OBESITY: string | null;
  }